azure_function/requirements.txt
# Azure Functions
azure-functions==1.18.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Azure Services
azure-storage-blob==12.19.0
azure-data-tables==12.4.0
azure-ai-formrecognizer==3.3.2
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.10.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
requests==2.31.0
